Use trailing debounce for subsequent filtering
Leading debounce was intended to keep the latency short for the first
suggestion popup, but it is not ideal for subsequent filtering.

If using server side fuzzy completion it won't take the characters typed
between firing off the request and receiving the result into
consideration. Switching to trailing debounce for the subsequent
filtering should ensure that if you stop typing it will always end up
narrowing the result.