Improve CompletionItem conversion handling for vscode-json-languageserver
In a file like:

```
{
  "|
}
```

The column retrieved via `vim.fn.match(line_to_cursor, '\\k*$')` will be
3

The language server has a different starting point and includes the `"`
in the completion items.

```
label: arrow_spacing
textEdit: table: 0x7fde68f19388
  newText: "arrow_spacing"
  range: table: 0x7fde68f193d0
    end: table: 0x7fde68f194d0
      character: 3
      line: 2
    start: table: 0x7fde68f19418
      character: 2
      line: 2
```

This starting point divergence must be accounted for when deciding if
the `label` or `newText` should be used as `word`.

If `label` would be used, typing any character would end the completion.
Candidates would stop matching because `complete()` was called with the
lower starting position _before_ the `"`.