Check for existing client when detaching
If detach is called when no client is attached, nvim_del_augroup_by_name
returns an error because the given augroup does not exist. And even if
that call did succeed, we would quickly hit an error when accessing `c`,
which is nil.

This case can happen when the LSP server fails to start. When that
happens, an LspDetach event is fired and if the user has an LspDetach
autocommand that calls nvim_lsp_compl.detach(), this error will occur.

The solution is simply to check that the client being detached actually
exists, and exit early if it does not.