Decouple signature/completion trigger/timer; use one autocmd per buffer
Instead of sharing one timer between signature and completion triggers
there are now two separate timers to avoid one canceling out the other.
This sometimes led to trigger characters not firing the completion when
it should have.

It also changes the multi-client handling a bit - instead of having a
autocmd per client and buffer the autocmds are now per buffer.